requirements.txt
numpy==1.21.0
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.2
onnx==1.9.0
onnxruntime==1.8.0
tvmai==0.8.0
pillow==8.3.1
matplotlib==3.4.2
scikit-learn==0.24.2
pytest==6.2.4
pycuda==2021.1
pyopencl==2021.2.10
Cython==0.29.24
